refactor(*): apply new recommended eslint rules
