+++ src/app/shared/model/color.model.ts
@@ -29,6 +29,7 @@ export class Color {
       try {
         color = JSON.parse(color);
       } catch (e) {
+        console.error(e);
         throw new Error(`Could not parse color (not a valid JSON): "${color}"`);
       }
     }
@@ -56,6 +57,7 @@ export class Color {
       try {
         shades.push(Shade.parse(shade));
       } catch (e) {
+        console.error(e);
         throw new Error(`Could not parse color (invalid shade): "${shade}"`);
       }
     }
