+++ src/app/shared/model/palette.model.ts
@@ -46,6 +46,7 @@ export class Palette {
       try {
         palette = JSON.parse(palette);
       } catch (e) {
+        console.error(e);
         throw new Error(`Could not parse palette (not a valid JSON): "${palette}"`);
       }
     }
@@ -77,6 +78,7 @@ export class Palette {
       try {
         colors.push(Color.parse(color));
       } catch (e) {
+        console.error(e);
         throw new Error(`Could not parse palette (invalid color): "${color}"`);
       }
     }
