+++ src/index.html
@@ -133,7 +133,6 @@
   </head>
 
   <body class="scroll-smooth bg-neutral-50 text-neutral-900 dark:bg-neutral-900 dark:text-neutral-50">
-    <!-- eslint-disable-next-line @angular-eslint/template/prefer-self-closing-tags -->
     <rp-root></rp-root>
     <noscript> Please enable JavaScript to continue using this application. </noscript>
   </body>
